+++ example/lib/example.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:grouped_list/grouped_list.dart';
 
-void main() => runApp(MyApp());
+void main() => runApp(const MyApp());
 
 List _elements = [
   {'name': 'John', 'group': 'Team A'},
@@ -13,6 +13,8 @@ List _elements = [
 ];
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -23,41 +25,44 @@ class MyApp extends StatelessWidget {
       ),
       home: Scaffold(
         appBar: AppBar(
-          title: Text('Grouped List View Example'),
+          title: const Text('Grouped List View Example'),
         ),
-        body: GroupedListView<dynamic, String>(
-          elements: _elements,
-          groupBy: (element) => element['group'],
-          groupComparator: (value1, value2) => value2.compareTo(value1),
-          itemComparator: (item1, item2) =>
-              item1['name'].compareTo(item2['name']),
-          order: GroupedListOrder.DESC,
-          useStickyGroupSeparators: true,
-          groupSeparatorBuilder: (String value) => Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: Text(
-              value,
-              textAlign: TextAlign.center,
-              style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-            ),
-          ),
-          itemBuilder: (c, element) {
-            return Card(
-              elevation: 8.0,
-              margin: EdgeInsets.symmetric(horizontal: 10.0, vertical: 6.0),
-              child: Container(
-                child: ListTile(
-                  contentPadding:
-                      EdgeInsets.symmetric(horizontal: 20.0, vertical: 10.0),
-                  leading: Icon(Icons.account_circle),
-                  title: Text(element['name']),
-                  trailing: Icon(Icons.arrow_forward),
-                ),
-              ),
-            );
-          },
+        body: _createGroupedListView(),
+      ),
+    );
+  }
+
+  _createGroupedListView() {
+    return GroupedListView<dynamic, String>(
+      elements: _elements,
+      groupBy: (element) => element['group'],
+      groupComparator: (value1, value2) => value2.compareTo(value1),
+      itemComparator: (item1, item2) => item1['name'].compareTo(item2['name']),
+      order: GroupedListOrder.DESC,
+      useStickyGroupSeparators: true,
+      groupSeparatorBuilder: (String value) => Padding(
+        padding: const EdgeInsets.all(8.0),
+        child: Text(
+          value,
+          textAlign: TextAlign.center,
+          style: const TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
         ),
       ),
+      itemBuilder: (c, element) {
+        return Card(
+          elevation: 8.0,
+          margin: const EdgeInsets.symmetric(horizontal: 10.0, vertical: 6.0),
+          child: SizedBox(
+            child: ListTile(
+              contentPadding:
+                  const EdgeInsets.symmetric(horizontal: 20.0, vertical: 10.0),
+              leading: const Icon(Icons.account_circle),
+              title: Text(element['name']),
+              trailing: const Icon(Icons.arrow_forward),
+            ),
+          ),
+        );
+      },
     );
   }
 }