+++ lib/src/grouped_list_order.dart
@@ -1,5 +1,6 @@
-import '../grouped_list.dart';
+import './../grouped_list.dart';
 import '../sliver_grouped_list.dart';
 
 /// Used to define the order of a [GroupedListView] or [SliverGroupedListView].
+// ignore: constant_identifier_names
 enum GroupedListOrder { ASC, DESC }